nlcs/backend/requirements.txt
fastapi
uvicorn
pymysql
sqlalchemy
python-multipart
pymupdf
pageindex
groq
python-dotenv
kagglehub
pandas
httpx
